refactor: delete sendErrorResponse function
