+++ src/controllers/handleGetRequest.ts
@@ -1,16 +1,4 @@
-import { IncomingMessage, ServerResponse } from 'http';
-import { BASE_URL } from '..';
+import { ServerResponse } from 'http';
 
-const requestHandlers = (
-  req: IncomingMessage,
-  res: ServerResponse,
-  method: string,
-  reqUrl: string
-) => {
-  const userId = reqUrl.substring(BASE_URL.length + 1);
-
-  if (method === 'GET' && userId) {
-  }
-};
-
-export default requestHandlers;
+const handleGetRequest = (res: ServerResponse, reqUrl: string) => {};
+export default handleGetRequest;
