+++ src/controllers/requestHandlers.ts
@@ -1,0 +1,17 @@
+import { IncomingMessage, ServerResponse } from 'http';
+import { BASE_URL } from '..';
+
+const requestHandlers = (
+  req: IncomingMessage,
+  res: ServerResponse,
+  method: string,
+  reqUrl: string
+) => {
+  const userId = reqUrl.substring(BASE_URL.length + 1);
+
+  if (method === 'GET' && !userId) {
+    res.write('users');
+  }
+};
+
+export default requestHandlers;
